Change minimumRssiThreshold type to integer

+++ BLE Indoor Positioning/src/main/java/com/nexenio/bleindoorpositioning/IndoorPositioning.java
@@ -34,7 +34,7 @@ public class IndoorPositioning implements LocationProvider, BeaconUpdateListener
     public static final double MAXIMUM_MOVEMENT_SPEED_NOT_SET = -1;
     private double maximumMovementSpeed = MAXIMUM_MOVEMENT_SPEED_NOT_SET;
     private double rootMeanSquareThreshold = ROOT_MEAN_SQUARE_THRESHOLD_STRICT;
-    private double minimumRssiThreshold = -70;
+    private int minimumRssiThreshold = -70;
 
     private static IndoorPositioning instance;
 
@@ -202,11 +202,11 @@ public class IndoorPositioning implements LocationProvider, BeaconUpdateListener
         return rootMeanSquareThreshold;
     }
 
-    public double getMinimumRssiThreshold() {
+    public int getMinimumRssiThreshold() {
         return minimumRssiThreshold;
     }
 
-    public void setMinimumRssiThreshold(double minimumRssiThreshold) {
+    public void setMinimumRssiThreshold(int minimumRssiThreshold) {
         this.minimumRssiThreshold = minimumRssiThreshold;
     }
 }
